Improve the graph visuals
Use C3.js to create the graph to get to a nicer result with less code.

Generalize the graph label definition so that multiple types of data can
be visualized (for example, temperature or light). The labels are taken
from the incoming data and code only assumes the data has keys starting
avg, min and max, but does not mandate the rest of the label or the
scale of the values.